// BE-Sgroup-2025-4-6-/src/models/inputTypeSchemas/CardMemberCountOrderByAggregateInputSchema.ts
import type { Prisma } from '@prisma/client';

import { z } from 'zod';
import { SortOrderSchema } from './SortOrderSchema';

export const CardMemberCountOrderByAggregateInputSchema: z.ZodType<Prisma.CardMemberCountOrderByAggregateInput> = z.strictObject({
  id: z.lazy(() => SortOrderSchema).optional(),
  cardId: z.lazy(() => SortOrderSchema).optional(),
  userId: z.lazy(() => SortOrderSchema).optional(),
  joinedAt: z.lazy(() => SortOrderSchema).optional(),
});

export default CardMemberCountOrderByAggregateInputSchema;